requirements.txt
Jinja2==3.1.2
parse==1.18.0
snowflake-connector-python==3.0.3
pyyaml==5.4.1
argparse==1.4.0
